+++ components/Banners/LandingBanner.tsx
@@ -6,7 +6,7 @@ import Title from '../Header/Title';
 import Description from '../Header/Description';  
 import { CylinderSVG } from '../SVGs';
 // UTILS
-import { cn } from '@/utils/helpers';
+import { cn } from '@/utils/helpers/client';
 
 export default async function LandingBanner() {
   return(
