+++ components/DashboardList/index.tsx
@@ -4,7 +4,7 @@
 import Link from "next/link"
 import { usePathname } from "next/navigation"
 // UTILS
-import { cn } from "@/utils/helpers"
+import { cn } from "@/utils/helpers/client"
 // TYPES
 import type { NavItem } from "@/types"
