+++ components/MDX/index.tsx
@@ -4,7 +4,7 @@ import Image from "next/image";
 // HOOKS
 import { useMDXComponent } from "next-contentlayer/hooks";
 // UTILS
-import { cn } from '@/utils/helpers';
+import { cn } from "@/utils/helpers/client";
 
 const components = {
   h1: ({ className, ...props}: { className?: string}) => (
